Changed "State" form input to a select
Also added a key => value array of state names and abbreviations. This
is what the select box iterates through. short code is starting to look
really ugly, and may need to be moved into a separate include file. If
I do this, I should also move the Admin Options panel code into a
separate include as well (for consistency).